+++ CHANGELOG.md
@@ -9,6 +9,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 * [BREAKING] Document#astNode and ParsedDocument#astNode are now an
   `AstNodeWithLanguage`, because we support inline documents in more than just
   HTML, as we've added an HTML-in-JS scanner.
+* `MultiUrlResolver` now delegates the `relative()` method to the first
+  `UrlResolver` in its `_resolvers` array that can `resolve()` the
+  destination URL.  Makes it possible now to rely on the Analyzer's
+  resolver to return a valid `PackageRelativeUrl` from a resolved URL.
 <!-- Add new, unreleased changes here. -->
 
 ## [3.0.0-pre.8] - 2017-01-18